+++ src/init-action.ts
@@ -36,13 +36,15 @@ import { Feature, featureConfig, Features } from "./feature-flags";
 import {
   checkInstallPython311,
   cleanupDatabaseClusterDirectory,
+  getOverlayDatabaseMode,
   initCodeQL,
   initConfig,
   runInit,
 } from "./init";
 import { Language } from "./languages";
 import { getActionsLogger, Logger } from "./logging";
-import { parseRepositoryNwo } from "./repository";
+import { OverlayDatabaseMode } from "./overlay-database-utils";
+import { getRepositoryNwo } from "./repository";
 import { ToolsSource } from "./setup-codeql";
 import {
   ActionName,
@@ -279,9 +281,7 @@ async function run() {
   checkGitHubVersionInRange(gitHubVersion, logger);
   checkActionVersion(getActionVersion(), gitHubVersion);
 
-  const repositoryNwo = parseRepositoryNwo(
-    getRequiredEnvParam("GITHUB_REPOSITORY"),
-  );
+  const repositoryNwo = getRepositoryNwo();
 
   const features = new Features(
     gitHubVersion,
@@ -395,7 +395,22 @@ async function run() {
   }
 
   try {
-    cleanupDatabaseClusterDirectory(config, logger);
+    const sourceRoot = path.resolve(
+      getRequiredEnvParam("GITHUB_WORKSPACE"),
+      getOptionalInput("source-root") || "",
+    );
+
+    const overlayDatabaseMode = await getOverlayDatabaseMode(
+      (await codeql.getVersion()).version,
+      config,
+      sourceRoot,
+      logger,
+    );
+    logger.info(`Using overlay database mode: ${overlayDatabaseMode}`);
+
+    if (overlayDatabaseMode !== OverlayDatabaseMode.Overlay) {
+      cleanupDatabaseClusterDirectory(config, logger);
+    }
 
     if (zstdAvailability) {
       await recordZstdAvailability(config, zstdAvailability);
@@ -675,11 +690,6 @@ async function run() {
       }
     }
 
-    const sourceRoot = path.resolve(
-      getRequiredEnvParam("GITHUB_WORKSPACE"),
-      getOptionalInput("source-root") || "",
-    );
-
     const tracerConfig = await runInit(
       codeql,
       config,
@@ -687,6 +697,7 @@ async function run() {
       "Runner.Worker.exe",
       getOptionalInput("registries"),
       apiDetails,
+      overlayDatabaseMode,
       logger,
     );
     if (tracerConfig !== undefined) {