+++ CHANGELOG.md
@@ -9,6 +9,19 @@ See the [releases page](https://github.com/github/codeql-action/releases) for th
   your `tools:` property to a specific version of the CodeQL CLI earlier than v2.20.4, you will need to update to at least CodeQL v2.20.4 or disable
   `actions` analysis.
 
+## 3.28.13 - 24 Mar 2025
+
+No user facing changes.
+
+## 3.28.12 - 19 Mar 2025
+
+- Dependency caching should now cache more dependencies for Java `build-mode: none` extractions. This should speed up workflows and avoid inconsistent alerts in some cases.
+- Update default CodeQL bundle version to 2.20.7. [#2810](https://github.com/github/codeql-action/pull/2810)
+
+## 3.28.11 - 07 Mar 2025
+
+- Update default CodeQL bundle version to 2.20.6. [#2793](https://github.com/github/codeql-action/pull/2793)
+
 ## 3.28.10 - 21 Feb 2025
 
 - Update default CodeQL bundle version to 2.20.5. [#2772](https://github.com/github/codeql-action/pull/2772)